+++ tests/test_metrics.py
@@ -42,16 +42,10 @@ def test_metrics_extractor(tmp_path):
         # MetricBase has lookup by commit
         if isinstance(metric, ChangeMetricBase):
 
-            # One is required
-            assert results.get("by-group") or results.get("by-file")
-            assert "0.0.1..0.0.11" in results.get("by-group") or ["0.0.1..0.0.11"]
-            assert "0.0.1..0.0.11" in results.get("by-file") or ["0.0.1..0.0.11"]
+            # Top level should be versions
+            assert results.get("0.0.1..0.0.11")
 
         elif isinstance(metric, MetricBase):
 
             # One is required
-            assert results.get("by-group") or results.get("by-file")
-
-            # ensure if no result type is provided, passes
-            assert "0.0.1" in results.get("by-group") or ["0.0.1"]
-            assert "0.0.1" in results.get("by-file") or ["0.0.1"]
+            assert results.get("0.0.1")